CoreStack tests
