+++ infra/jest.config.ts
@@ -2,13 +2,14 @@ import type { Config } from 'jest';
 
 const config: Config = {
     coverageReporters: ['json'],
-    roots: ['<rootDir>/__tests__'],
-    setupFiles: ['<rootDir>/__tests__/setupTests.ts'],
+    roots: ['<rootDir>/tests'],
+    setupFiles: ['<rootDir>/tests/setupTests.ts'],
     testEnvironment: 'node',
     testMatch: ['**/*.test.ts'],
     transform: {
         '^.+\\.tsx?$': 'ts-jest',
     },
+    testPathIgnorePatterns: ['<rootDir>/tests/__faker__/'],
 };
 
 export default config;
